requirements/developments.txt
asgiref==3.8.1
autopep8==2.1.0
dj-database-url==2.1.0
Django==5.0.3
django-autoslug==1.9.9
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==24.2
django-silk==5.1.0
django-versatileimagefield==3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gprof2dot==2022.7.29
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
